Revert "LPS-122562 Move these announcements and calendar portlet preference upgrades to portal and rename since they must run before the new upgrade. Also move over the announcements portletId upgrade so that it's always run before the preferences one. These should have all been run upgrading to 7.1.x. See LPS-55278, LPS-60361 and LPS-84128."
This reverts commit f93b24089a94bba90bf868452f638eb0e4c03399.
