+++ modules/apps/announcements/announcements-web/src/main/java/com/liferay/announcements/web/internal/upgrade/AnnouncementsWebUpgrade.java
@@ -15,8 +15,11 @@
 package com.liferay.announcements.web.internal.upgrade;
 
 import com.liferay.announcements.web.internal.upgrade.v1_0_2.UpgradePermission;
+import com.liferay.announcements.web.internal.upgrade.v1_0_4.UpgradePortletPreferences;
+import com.liferay.portal.kernel.upgrade.BaseReplacePortletId;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.kernel.upgrade.UpgradeException;
+import com.liferay.portal.kernel.upgrade.UpgradeStep;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 import com.liferay.portal.upgrade.release.BaseUpgradeWebModuleRelease;
 
@@ -57,19 +60,44 @@ public class AnnouncementsWebUpgrade implements UpgradeStepRegistrator {
 
 		registry.register("0.0.0", "2.0.0", new DummyUpgradeStep());
 
-		registry.register("0.0.1", "1.0.1", new DummyUpgradeStep());
+		UpgradeStep upgradePortletId = new BaseReplacePortletId() {
+
+			@Override
+			protected String[][] getRenamePortletIdsArray() {
+				return new String[][] {
+					{
+						"1_WAR_soannouncementsportlet",
+						"com_liferay_announcements_web_portlet_" +
+							"AnnouncementsPortlet"
+					},
+					{
+						"83",
+						"com_liferay_announcements_web_portlet_AlertsPortlet"
+					},
+					{
+						"84",
+						"com_liferay_announcements_web_portlet_" +
+							"AnnouncementsPortlet"
+					}
+				};
+			}
+
+		};
+
+		registry.register("0.0.1", "1.0.1", upgradePortletId);
 
 		// See LPS-65946
 
-		registry.register("1.0.0", "1.0.1", new DummyUpgradeStep());
+		registry.register("1.0.0", "1.0.1", upgradePortletId);
 
 		registry.register("1.0.1", "1.0.2", new UpgradePermission(true));
 
 		// See LPS-69656
 
-		registry.register("1.0.2", "1.0.3", new UpgradePermission(true));
+		registry.register(
+			"1.0.2", "1.0.3", upgradePortletId, new UpgradePermission(true));
 
-		registry.register("1.0.3", "1.0.4", new DummyUpgradeStep());
+		registry.register("1.0.3", "1.0.4", new UpgradePortletPreferences());
 
 		registry.register(
 			"1.0.4", "2.0.0",
