+++ portal-impl/src/com/liferay/portal/upgrade/v7_1_x/PortalUpgradeProcessRegistryImpl.java
@@ -51,20 +51,6 @@ public class PortalUpgradeProcessRegistryImpl
 			new Version(2, 0, 3), new UpgradePortalPreferences());
 
 		upgradeProcesses.put(new Version(2, 0, 4), new UpgradeUserGroup());
-
-		upgradeProcesses.put(
-			new Version(2, 0, 5), new UpgradeAnnouncementsPortletId());
-
-		upgradeProcesses.put(
-			new Version(2, 0, 6), new UpgradeAnnouncementsPortletPreferences());
-
-		upgradeProcesses.put(
-			new Version(2, 0, 7),
-			new UpgradeCalendarTimeFormatPortletPreferences());
-
-		upgradeProcesses.put(
-			new Version(2, 0, 8),
-			new UpgradeCalendarClassNameIdsPortletPreferences());
 	}
 
 }
